Classification/requirements.txt
tensorflow==2.20.0
numpy>=1.23
pillow>=9.5
scikit-learn>=1.2
